services/ai-service/requirements.txt
# ============================================
# NEURAXIS AI Service - Python Dependencies
# ============================================

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
httpx==0.26.0

# AI/ML
torch>=2.2.0
transformers>=4.36.2
Pillow>=10.2.0
numpy>=1.26.3
scikit-learn>=1.4.0

# Medical Imaging
pydicom==2.4.4

# Monitoring & Logging
structlog==24.1.0
sentry-sdk[fastapi]==1.39.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code Quality
ruff==0.1.13
mypy==1.8.0

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0.post1
